测试通过
新改为
ApplicationContext applicationContext = new ClassPathXmlApplicationContext("classpath:spring/applicationContext-*.xml");
		CyclePicService bean = applicationContext.getBean(CyclePicService.class);
		System.out.println(bean.getid().toString());;

+++ companyinfo/src/main/java/pojo/CiCyclePic.java
@@ -3,7 +3,7 @@ package pojo;
 import java.util.Date;
 
 public class CiCyclePic {
-    private int id;
+    private Byte id;
 
     private String name;
 
@@ -11,11 +11,11 @@ public class CiCyclePic {
 
     private Date updateTime;
 
-    public int getId() {
+    public Byte getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Byte id) {
         this.id = id;
     }
